CookieSession: share common settings with SessionCookieConfig.
Server session and client based session share common settings like: name, path, httpOnly, secure, expires with SessionCookieConfig.
Client based session specific settings (like secureToken, maxInactivityTime) are left in ClientSideSessionConfig.
